@@ -8,7 +8,7 @@
     "lint": "pnpm run lint:js:all --fix && pnpm run lint:css:all --fix",
     "lint:css": "stylelint --max-warnings 0 **/*.{css,vue}",
     "lint:css:all": "pnpm run lint:css **/*.{css,vue}",
-    "lint:js": "eslint --max-warnings 0 --ext .js,.ts,.vue",
+    "lint:js": "eslint --max-warnings 0",
     "lint:js:all": "pnpm run lint:js .",
     "prepare": "husky install .husky",
     "serve": "vite preview",
@@ -30,8 +30,6 @@
     "@iconify/json": "2.2.198",
     "@mdi/js": "7.4.47",
     "@types/node": "20.12.5",
-    "@typescript-eslint/eslint-plugin": "7.5.0",
-    "@typescript-eslint/parser": "7.5.0",
     "@unocss/core": "0.58.9",
     "@unocss/preset-attributify": "0.58.9",
     "@unocss/preset-icons": "0.58.9",
@@ -40,15 +38,15 @@
     "@unocss/transformer-directives": "0.58.9",
     "@unocss/transformer-variant-group": "0.58.9",
     "@vitejs/plugin-vue": "5.0.4",
-    "@vue-macros/reactivity-transform": "0.4.4",
     "@vue/compiler-sfc": "3.4.21",
     "@vue/runtime-core": "3.4.21",
     "@vue/test-utils": "2.4.5",
-    "eslint": "8.57.0",
+    "eslint": "9.0.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-prettier-vue": "5.0.0",
     "eslint-plugin-simple-import-sort": "12.0.0",
     "eslint-plugin-vue": "9.24.0",
+    "espree": "^10.0.1",
     "husky": "9.0.11",
     "jsdom": "24.0.0",
     "lint-staged": "15.2.2",
@@ -59,6 +57,7 @@
     "stylelint-config-idiomatic-order": "10.0.0",
     "stylelint-config-standard": "36.0.0",
     "typescript": "5.4.4",
+    "typescript-eslint": "7.5.0",
     "unocss": "0.58.9",
     "unocss-preset-daisy": "7.0.0",
     "unocss-preset-scrollbar": "0.3.1",
@@ -88,5 +87,13 @@
   "packageManager": "pnpm@8.15.6",
   "prettier": {
     "singleAttributePerLine": true
+  },
+  "type": "module",
+  "pnpm": {
+    "peerDependencyRules": {
+      "allowedVersions": {
+        "eslint": "9"
+      }
+    }
   }
 }
